Test Chrome Manager

+++ .platform/hooks/predeploy/01_chrome.sh
@@ -1,11 +1,29 @@
 #!/bin/bash
 
-# Install Google Chrome
-curl -sS https://intoli.com/install-google-chrome.sh | bash || exit 1
-mv -f /usr/bin/google-chrome-stable /usr/bin/google-chrome || exit 1
-google-chrome --version && which google-chrome || exit 1
+# Check if Google Chrome is already installed
+if ! command -v google-chrome &> /dev/null
+then
+    # Install Google Chrome
+    curl -sS https://intoli.com/install-google-chrome.sh | bash
+    if [ $? -ne 0 ]; then
+        echo "Failed to install Google Chrome"
+        exit 1
+    fi
+    mv -f /usr/bin/google-chrome-stable /usr/bin/google-chrome
+    if [ $? -ne 0 ]; then
+        echo "Failed to move Google Chrome executable"
+        exit 1
+    fi
+else
+    echo "Google Chrome is already installed"
+fi
 
-# Clean up temporary files
-rm chromedriver_linux64.zip
+# Print the version and location
+google-chrome --version && which google-chrome
 
-echo "Google Chrome installation completed successfully."
+# Check if the temporary file exists before trying to remove it
+if [ -f "chromedriver_linux64.zip" ]; then
+    rm chromedriver_linux64.zip
+fi
+
+echo "Google Chrome installation script completed successfully."
